Move drawerWidth to new constants file. Create mobileCutoff constant
