+++ src/components/common/constants.js
@@ -1,4 +1,7 @@
 // The width of the navigation drawer
 // This constant is used both within Nav and MainPageContainer
 // Unfortunately, we can't just use LESS because it must be set in the antd Drawer component as a prop
-export default 256;
+export const drawerWidth = 256;
+
+// The width in pixels where most pages shift from desktop to mobile view
+export const mobileCutoff = 1000;
